% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agilent.r
\docType{package}
\name{agilent-package}
\alias{agilent}
\alias{agilent-package}
\title{Reading and normalizing Agilent one color microarray data files.}
\description{
Functions for reading, exploring and normalizing Agilent one color microarray data.
}
\author{
David Montaner \email{dmontaner@cipf.es}
}

