% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreg.r
\name{flags}
\alias{flags}
\title{Extract flagout form an ExpressionSet.}
\usage{
flags(object, notfound2zero = TRUE, verbose = FALSE)
}
\arguments{
\item{object}{an Agilent One Color ExpressionSet.}

\item{notfound2zero}{If TRUE and flagout matrix is not found in the assayData a zero matrix is returned.}

\item{verbose}{verbose.}
}
\value{
B matrix from the Agilent ExpressionSet
}
\description{
Retrieve flagout information data from Agilent One Color ExpressionSet.
}
\details{
This function tries to access the flagout matrix in the assayData of an Agilent One Color ExpressionSet.
If notfound2zero is FALSE, when B matrix is not found in the assayData, a NULL value is returned.
}
\author{
David Montaner \email{dmontaner@cipf.es}
}
\seealso{
\code{\link{foreg}} \code{\link{backg}} \code{\link{readAgilentHeader}} \code{\link{readGPR}}
}
\keyword{flags}

