% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAgilentHeaderS.r
\name{readAgilentHeader}
\alias{readAgilentHeader}
\title{Read Agilent header form a single file.}
\usage{
readAgilentHeader(file)
}
\arguments{
\item{file}{name of the file which header is to be read.}
}
\value{
A list of 3 components containing the FEPARAMS, STATS, and FEATURE TYPE FROM the header of the Agilent file.
Each element of the list is a list itself.
}
\description{
Reads Agilent header from a unique txt exported file.
}
\details{
readAgilentHeaderS should be generally used.
Reads Agilent header from one or several txt exported files.
}
\examples{
myfile <- file.path (system.file ("exampledata", package = "agilent"), "agilent1.txt")
myfile
readAgilentHeader (file = myfile)

}
\seealso{
\code{\link{readAgilentHeaderS}} \code{\link{readAgilent}} \code{\link{readGPR}}
}
\author{
David Montaner \email{dmontaner@cipf.es}
}
\keyword{agilent}
\keyword{header}
\keyword{read}
